Extract some promise util functions
