+++ lib/plugins/bed.js
@@ -1,15 +1,9 @@
 const Vec3 = require('vec3').Vec3
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
 const CARDINAL_DIRECTIONS = ['south', 'west', 'north', 'east']
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
-}
 
 function inject (bot) {
   bot.isSleeping = false
