+++ lib/plugins/blocks.js
@@ -2,6 +2,7 @@ const Vec3 = require('vec3').Vec3
 const assert = require('assert')
 const Painting = require('../painting')
 const Location = require('../location')
+const { callbackify, sleep } = require('../promise_utils')
 
 const { OctahedronIterator } = require('../iterators')
 
@@ -116,7 +117,7 @@ function inject (bot, { version, storageBuilder }) {
       for (let y = -2; y <= 2; y++) {
         if (bot.world.getColumnAt(bot.entity.position.plus(new Vec3(x, 0, y).scaled(16))) == null) {
           // keep wait
-          await new Promise(resolve => setTimeout(resolve, 100))
+          await sleep(100)
           await waitForChunksToLoad()
         }
       }
@@ -503,11 +504,4 @@ function onesInShort (n) {
     count = ((1 << i) & n) ? count + 1 : count
   }
   return count
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
