+++ lib/plugins/book.js
@@ -1,14 +1,8 @@
 const assert = require('assert')
 const { once } = require('events')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
-}
 
 function inject (bot, { version }) {
   const mcData = require('minecraft-data')(version)
