+++ lib/plugins/chest.js
@@ -1,5 +1,6 @@
 const Chest = require('../chest')
 const assert = require('assert')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -55,11 +56,4 @@ function inject (bot, { version }) {
   }
 
   bot.openChest = openChest
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
