+++ lib/plugins/craft.js
@@ -1,5 +1,6 @@
 const assert = require('assert')
 const { once } = require('events')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -236,11 +237,4 @@ function inject (bot, { version }) {
   bot.craft = callbackify(craft)
   bot.recipesFor = recipesFor
   bot.recipesAll = recipesAll
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
