+++ lib/plugins/creative.js
@@ -1,5 +1,6 @@
 const assert = require('assert')
 const Vec3 = require('vec3').Vec3
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -86,11 +87,4 @@ function inject (bot, { version }) {
 // this should be in the vector library
 function vecMagnitude (vec) {
   return Math.sqrt(vec.x * vec.x + vec.y * vec.y + vec.z * vec.z)
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
