+++ lib/plugins/dispenser.js
@@ -1,5 +1,6 @@
 const assert = require('assert')
 const Dispenser = require('../dispenser')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -40,11 +41,4 @@ function inject (bot) {
   }
 
   bot.openDispenser = openDispenser
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
