+++ lib/plugins/enchantment_table.js
@@ -1,6 +1,7 @@
 const EnchantmentTable = require('../enchantment_table')
 const assert = require('assert')
 const { once } = require('events')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -76,11 +77,4 @@ function inject (bot, { version }) {
   }
 
   bot.openEnchantmentTable = openEnchantmentTable
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
