+++ lib/plugins/furnace.js
@@ -1,5 +1,6 @@
 const Furnace = require('../furnace')
 const assert = require('assert')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
@@ -98,11 +99,4 @@ function inject (bot, { version }) {
   }
 
   bot.openFurnace = openFurnace
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
