+++ lib/plugins/inventory.js
@@ -1,6 +1,7 @@
 const assert = require('assert')
 const Vec3 = require('vec3').Vec3
 const { once } = require('events')
+const { callbackify, sleep } = require('../promise_utils')
 
 module.exports = inject
 
@@ -504,14 +505,9 @@ function inject (bot, { version, hideErrors }) {
     // if you click on the quick bar and have dug recently,
     // wait a bit
     if (slot >= bot.QUICK_BAR_START && bot.lastDigTime != null) {
-      const timeSinceLastDig = new Date() - bot.lastDigTime
-      if (timeSinceLastDig < DIG_CLICK_TIMEOUT) {
-        return new Promise(resolve => {
-          setTimeout(async () => {
-            await clickWindow(slot, mouseButton, mode)
-            resolve()
-          }, DIG_CLICK_TIMEOUT - timeSinceLastDig)
-        })
+      let timeSinceLastDig
+      while ((timeSinceLastDig = new Date() - bot.lastDigTime) < DIG_CLICK_TIMEOUT) {
+        await sleep(DIG_CLICK_TIMEOUT - timeSinceLastDig)
       }
     }
     const window = bot.currentWindow || bot.inventory
@@ -538,21 +534,18 @@ function inject (bot, { version, hideErrors }) {
       item: Item.toNotch(click.item)
     })
 
-    return new Promise((resolve, reject) => {
-      bot.once(`confirmTransaction${actionId}`, (success) => {
-        if (success) {
-          resolve()
-        } else {
-          reject(new Error('Server rejected transaction.'))
-        }
-      })
+    const response = once(bot, `confirmTransaction${actionId}`)
 
-      // notchian servers are assholes and only confirm certain transactions.
-      if (!window.transactionRequiresConfirmation(click)) {
-        // jump the gun and accept the click
-        confirmTransaction(window.id, actionId, true)
-      }
-    })
+    // notchian servers are assholes and only confirm certain transactions.
+    if (!window.transactionRequiresConfirmation(click)) {
+      // jump the gun and accept the click
+      confirmTransaction(window.id, actionId, true)
+    }
+
+    const [success] = await response
+    if (!success) {
+      throw new Error('Server rejected transaction.')
+    }
   }
 
   async function putAway (slot) {
@@ -676,13 +669,6 @@ function inject (bot, { version, hideErrors }) {
   bot.openEntity = openEntity
   bot.moveSlotItem = callbackify(moveSlotItem)
   bot.updateHeldItem = updateHeldItem
-}
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(undefined, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
 }
 
 function vectorToDirection (v) {
