+++ lib/plugins/physics.js
@@ -3,6 +3,7 @@ const assert = require('assert')
 const math = require('../math')
 const conv = require('../conversions')
 const { performance } = require('perf_hooks')
+const { callbackify } = require('../promise_utils')
 
 const { Physics, PlayerState } = require('prismarine-physics')
 
@@ -196,13 +197,6 @@ function inject (bot) {
 
   function noop (err) {
     if (err) throw err
-  }
-
-  function callbackify (f) {
-    return function (...args) {
-      const cb = args[f.length]
-      return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-    }
   }
 
   let lookAtCb = noop
