+++ lib/plugins/simple_inventory.js
@@ -1,16 +1,10 @@
 const assert = require('assert')
+const { callbackify } = require('../promise_utils')
 
 module.exports = inject
 
 const QUICK_BAR_COUNT = 9
 const QUICK_BAR_START = 36
-
-function callbackify (f) {
-  return function (...args) {
-    const cb = args[f.length]
-    return f(...args).then(r => { if (cb) { cb(null, r) } return r }, err => { if (cb) { cb(err) } else throw err })
-  }
-}
 
 function inject (bot) {
   let nextQuickBarSlot = 0
